internal/cmdtest: Move test geth process runner into package
